Restructure again to put all logic in datamatic library
