+++ test/test_smoketest.py
@@ -1,2 +1,4 @@
+from datamatic import api
+
 def test_foo():
     assert 2 == 2
